Move MessagePopup to dialog module
