+++ src/ui/dialog/command.rs
@@ -22,7 +22,7 @@ use crate::ComponentInputResult;
 use crate::commander::new_commander;
 use crate::ui::Component;
 use crate::ui::ComponentAction;
-use crate::ui::message_popup::MessagePopup;
+use crate::ui::dialog::MessagePopup;
 use crate::ui::utils::centered_rect_line_height;
 
 pub struct CommandPopup<'a> {
