+++ src/ui/dialog/mod.rs
@@ -10,7 +10,9 @@ until it is closed.
 mod bookmark_set;
 mod command;
 mod help;
+mod message;
 
 pub use bookmark_set::BookmarkSetPopup;
 pub use command::CommandPopup;
 pub use help::HelpPopup;
+pub use message::MessagePopup;
